Refactor tracking state
