+++ projects/make-a-vis/src/lib/tracking-popup/tracking-popup.component.scss
