+++ projects/make-a-vis/src/public_api.ts
@@ -3,3 +3,6 @@
  */
 export * from './lib/make-a-vis.component';
 export * from './lib/make-a-vis.module';
+export * from './lib/tracking-popup/tracking-popup.component';
+export * from './lib/tracking-popup/tracking-popup.module';
+export * from './lib/shared/store/tracking-state';
