+++ src/app/app.module.ts
@@ -11,12 +11,8 @@ import { AppComponent } from './app.component';
 import { AppUpdateNotificationComponent } from './app-update-notification/app-update-notification.component';
 
 import { MatSnackBarModule } from '@angular/material/snack-bar';
-import { TrackingPopupModule } from './tracking-popup/tracking-popup.module';
+import { TrackingPopupModule, INITIAL_TELEMETRY_SETTING } from 'make-a-vis';
 import { NgxGoogleAnalyticsModule, NgxGoogleAnalyticsRouterModule } from 'ngx-google-analytics';
-import { INITIAL_TELEMETRY_SETTING } from './services/tracking-state';
-import { NgxsDataPluginModule } from '@ngxs-labs/data';
-import { NgxsModule } from '@ngxs/store';
-import { TrackingState } from './services/tracking-state';
 
 const appRoutes: Routes = [
   { path: '', component: AppComponent }
@@ -38,12 +34,7 @@ const appRoutes: Routes = [
     MatSnackBarModule,
     TrackingPopupModule,
     NgxGoogleAnalyticsModule.forRoot(INITIAL_TELEMETRY_SETTING === false ? '' : environment.googleAnalyticsTag),
-    NgxGoogleAnalyticsRouterModule,
-    NgxsDataPluginModule.forRoot(),
-
-    NgxsModule.forRoot([TrackingState], {
-      developmentMode: !environment.production
-    })
+    NgxGoogleAnalyticsRouterModule
   ],
   providers: [],
   bootstrap: [AppComponent],
